% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_tree_risk.R
\name{classify_tree_risk}
\alias{classify_tree_risk}
\title{Classify tree drought risk}
\usage{
classify_tree_risk(spp, AET_PET)
}
\arguments{
\item{spp}{Species code(s) for which you want risk ratings.  Can pass multiple species using, for example \code{c("Sx","Pl")}.}

\item{AET_PET}{Vector of AET:PET ratios.  These will be cut into risk ratings.}
}
\value{
Data frame with column(s) named for tree species, and risk categories assigned to each AET:PET value.
}
\description{
Use this function to classify stand-level drought risk based on AET:PET ratio and tree species.
}
\details{
{
This function calls the \emph{tree_risk} data in this package, and uses species specific values to cut AET:PET into different drought risk.  The function cuts sites into one of four categories: Low (L), Medium (M), High (H) and Very High (VH).  A \emph{Very High} rating
indicates that the site is drier than observed distribution of the species.  Categories are assigned using
\emph{less than or equal to} the specific category rating.  For example, if the lowest AET:PET threshold for a species is 0.81, then AET:PET values of .81 or lower will be categorized as having
a \strong{Very High} drought risk rating.

The
\emph{tree_risk} data will be updated as research further refines species-specific AET:PET thresholds.
\cr
\cr
Species codes must match those found in \code{levels(tree_risk$Species)}.

}
}
\examples{
classify_tree_risk(spp=c("Pl","Sx","Fd"),AET_PET=c(0.45,0.78,0.88,0.99))

# Can use 'mutate' to add columns to dataframe


data.frame(AET_PET=c(0.45,0.78,0.88,0.99)) \%>\%
dplyr::mutate(classify_tree_risk(spp=c("Pl","At"),AET_PET=AET_PET))

}
