% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_risk.R
\docType{data}
\name{tree_risk}
\alias{tree_risk}
\title{Drought tolerance ratings developed for tree species in British Columbia}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 33 rows and 3 columns.
}
\usage{
tree_risk
}
\description{
Dataset of estimated species AET:PET thresholds for several tree species in BC.  These data are combined with stand-level AET:PET to estimate drought risk,
using the \link{classify_tree_risk} function in this package.
}
\details{
{

The data frame consists of the following columns:

\strong{Tree species codes: }
The following species codes are listed in \emph{Species} column:
\enumerate{
\item Sx - hybrid spruce (\emph{Picea glauca} x \emph{P. engelmannii})
\item Bl - subalpine fir (\emph{Abies lasiocarpa})
\item Cw - western redcedar (\emph{Thuja plicata})
\item Hw - western hemlock (\emph{Tsuga heterophylla})
\item Fd - Interior Douglas-fir (\emph{Pseudotsuga menziesii} var. \emph{glauca})
\item Ep - Paper birch (\emph{Betula papyrifera})
\item At - Trembling aspen (\emph{Populus tremuloides})
\item Ac - Black cottonwood (\emph{Populus trichocarpa})
}
\cr

\strong{risk_cat:} Risk category.  \emph{VH} = very high and refers to sites with lower AET:PET than associated value in \strong{aet_pet_threshold} column. \emph{H} = high risk, and
\emph{M} = moderate risk.  Sites with AET:PET greater than threshold for \emph{M} sites are considered to have \emph{L} (low) drought risk.
}
}
\keyword{datasets}
